Refactor: Library crate.
Moved core functionality into a new library crate.  This is to make way for
integration testing.
